chimera-scraper/pyproject.toml
[tool.poetry]
name = "chimera-scraper"
version = "0.1.0"
description = "Advanced web scraper for review aggregation"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.24.1"
playwright = "^1.36.0"
pydantic = "^1.10.9"
loguru = "^0.7.0"
redis = "^4.5.5"
beautifulsoup4 = "^4.12.2"
python-dotenv = "^1.0.0"
tenacity = "^8.2.2"
aiofiles = "^23.1.0"
pandas = "^2.0.3"
lxml = "^4.9.2"
curl-cffi = "^0.5.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
